moves control panel to extra file to signal scope
