+++ gui/src/lib.rs
@@ -1,6 +1,26 @@
-pub use app::SecCardGameApp;
+use game_lib::world::game::Game;
 
 mod app;
 mod card_view_model;
 mod card_window;
 pub mod start;
+mod control_panel;
+
+pub struct SecCardGameApp {
+    game: Game,
+    input: Input,
+}
+
+enum Message {
+    Success(String),
+    Failure(String),
+    Warning(String),
+    None,
+}
+
+struct Input {
+    next_res: String,
+    pay_res: String,
+    message: Message,
+    multiplier: String,
+}
